requirements.txt
streamlit==1.52.2
pandas==2.3.3
numpy==2.4.0
matplotlib==3.10.0
seaborn==0.13.2

gspread==6.1.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0